Cleanup: - Remove now un-used code (WebsocketManager, and some things in VoiceChatBridge). - Split ConnectionAuthenticator into 3 classes to keep things clean.
TODO: Test
