+++ core/src/main/java/io/github/theodoremeyer/simplevoicegeyser/core/server/connection/ConnectionManager.java
@@ -59,7 +59,6 @@ public final class ConnectionManager {
 
         SvgConnection connection =
                 new SvgConnection(
-                        uuid,
                         session,
                         player
                 );
